@@ -2,6 +2,7 @@ package com.personal.portfolio.model;
 
 import java.util.*;
 
+import jakarta.validation.constraints.Email;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 import org.springframework.security.core.GrantedAuthority;
@@ -10,20 +11,18 @@ import org.springframework.security.core.userdetails.UserDetails;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Table(name = "users")
+/**
+ * Represents a user entity for the application, including authentication and
+ * authorization details. Implements {@link UserDetails} to integrate with
+ * Spring Security.
+ */
 @Entity
+@Table(name = "users")
 @Data
 @NoArgsConstructor
 public class User implements UserDetails {
@@ -33,65 +32,80 @@ public class User implements UserDetails {
 	@Column(nullable = false)
 	private Long id;
 
-	@Column(nullable = false, length = 100)
+	// The full name of the user.
 	private String fullName;
 
-	@Column(unique = true, length = 100, nullable = false)
+	// Unique email address for the user (used as the username in authentication).
+	@Column(unique = true, nullable = false)
+	@Email
 	private String email;
 
-	@Column(nullable = false)
+	// Encrypted password (excluded from serialization for security).
 	@JsonIgnore
 	private String password;
 
+	// Role of the user for authorization (e.g., ADMIN, USER).
 	@Enumerated(EnumType.STRING)
-	@Column(name = "role")
 	private Role role = Role.USER;
 
-	@Column(name = "enabled")
+	// Flags to control account status and credentials validity.
 	private boolean enabled = true;
-
-	@Column(name = "locked")
 	private boolean locked = false;
-
-	@Column(name = "account_non_expired")
 	private boolean accountNonExpired = true;
-
-	@Column(name = "credentials_non_expired")
 	private boolean credentialsNonExpired = true;
 
+	// Automatically managed timestamps for record creation and updates.
 	@CreationTimestamp
-	@Column(updatable = false, name = "created_at")
+	@Column(updatable = false)
 	private Date createdAt;
 
 	@UpdateTimestamp
-	@Column(name = "updated_at")
 	private Date updatedAt;
 
+	/**
+	 * Returns the email address of the user as the username for authentication.
+	 */
 	@Override
 	public String getUsername() {
 		return email;
 	}
 
+	/**
+	 * Returns whether the user is currently enabled.
+	 */
 	@Override
 	public boolean isEnabled() {
-		return this.enabled;
+		return enabled;
 	}
 
+	/**
+	 * Returns whether the user's account is not locked.
+	 */
 	@Override
 	public boolean isAccountNonLocked() {
-		return !this.locked;
+		return !locked;
 	}
 
+	/**
+	 * Returns whether the user's account is not expired.
+	 */
 	@Override
 	public boolean isAccountNonExpired() {
-		return this.accountNonExpired;
+		return accountNonExpired;
 	}
 
+	/**
+	 * Returns the authorities granted to the user.
+	 * This is based on the user's role.
+	 */
 	@Override
 	public Collection<? extends GrantedAuthority> getAuthorities() {
 		return Collections.singleton(new SimpleGrantedAuthority(role.name()));
 	}
 
+	/**
+	 * Overrides equals to compare users by their unique ID.
+	 */
 	@Override
 	public boolean equals(Object o) {
 		if (this == o) return true;
@@ -99,6 +113,9 @@ public class User implements UserDetails {
 		return id != null && id.equals(user.id);
 	}
 
+	/**
+	 * Overrides hashCode to provide a consistent hash based on the user's class.
+	 */
 	@Override
 	public int hashCode() {
 		return getClass().hashCode();
