@@ -3,11 +3,18 @@ package com.personal.portfolio;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest(classes = PortfolioApplication.class)
+/**
+ * Test suite for the Portfolio application.
+ * Ensures the application context loads correctly.
+ */
+@SpringBootTest
 class PortfolioApplicationTests {
 
+	/**
+	 * Verifies that the application context loads without issues.
+	 */
 	@Test
 	void contextLoads() {
+		// This test will pass if the application context is loaded successfully.
 	}
-
 }
